kk-plates/pyproject.toml
[project]
name = "kk-plates"
version = "1.0.0"
description = "Real-time plate counting and classification system for Kichi-Kichi restaurants"
authors = [{name = "KK Engineering", email = "engineering@kichi-kichi.com"}]
requires-python = ">=3.11"
dependencies = [
    "ultralytics>=8.1.0",
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "opencv-python>=4.8.0",
    "PyGObject>=3.44.0",
    "numpy>=1.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "typer[all]>=0.9.0",
    "structlog>=23.2.0",
    "requests>=2.31.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.7.0",
    "pyyaml>=6.0",
    "Pillow>=10.0.0",
    "onnx>=1.15.0",
    "onnxruntime>=1.16.0",
    "lap>=0.4.0",
    "prometheus-client>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "ipykernel>=6.25.0",
    "jupyter>=1.0.0",
]

[project.scripts]
kkplates = "kkplates.cli:app"

[build-system]
requires = ["setuptools>=67.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = {find = {where = ["src"]}}

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "W", "I", "N", "UP", "ANN", "B", "A", "C4", "SIM", "ARG"]
ignore = ["ANN101", "ANN102"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true